Opcion listar libros y autores registrados completados

+++ src/main/java/com/aluracursos/literalura/model/DatosLibros.java
@@ -11,7 +11,7 @@ public record DatosLibros(
         @JsonAlias("title") String titulo,
         @JsonAlias("authors") List<DatosAutores> autores,
         @JsonAlias("languages") List<String> idiomas,
-        @JsonAlias("download_count") Double numeroDescargas
+        @JsonAlias("download_count") Integer numeroDescargas
 
     ) {
 }
